base: Fix header images padding on content padding
This is more a workaround than an actual fix and it's for the header images from omnistyle.
This make it ignore the padding made with rounded_corner_content_padding on the main QS and
it applies to where the most visible change is done: On the QS icon part. It may break padding
on other places of qs, didn't saw anything of that.

This commit removes the padding of this example:
https://i.imgur.com/YzZtpxv.jpg